Fix Compiler Warnings in FileDialog.cpp

+++ Anim-Paint/include/Dialogs/FileDialog.hpp
@@ -77,7 +77,7 @@ public:
 	sf::Vector2i getSize();
 	void setPosition(sf::Vector2i position);
 	sf::Vector2i getPosition();
-	void setRange(float minX, float maxX);
+	void setRange(int minX, int maxX);
 
 	void cursorHover();
 	void handleEvent(const sf::Event& event);
